Add language-aware post routing
