--- src/utils/getPostImage.ts
@@ -1,30 +1,0 @@
-import fs from "fs";
-import path from "path";
-import matter from "gray-matter";
-import { PostMeta } from "@/types";
-
-export const getPost = async (slug: string): Promise<PostMeta> => {
-  const postsDirectory = path.join(process.cwd(), "src", "posts");
-  const fileContents = fs.readFileSync(path.join(postsDirectory, `${slug}.md`), "utf8");
-  const { data } = matter(fileContents);
-
-  return { ...data, slug } as PostMeta;
-};
-
-export const getImageUrl = async (slug: string): Promise<string> => {
-  const post = await getPost(slug);
-  const imagePath = post.image;
-
-  const normalizedImagePath = imagePath.startsWith('/') ? imagePath : `/${imagePath}`;
-
-  return normalizedImagePath;
-};
-
-export const parseDateString = (dateString: string | undefined): Date => {
-  if (!dateString || typeof dateString !== "string" || !dateString.includes("/")) {
-    // Return a very old date to push invalid/missing dates to the end
-    return new Date(0);
-  }
-  const [day, month, year] = dateString.split("/").map(Number);
-  return new Date(year, month - 1, day);
-};
